feat: complete C engine wasm build with modern web UI and 16-player network prototype

+++ include/PR/os.h
@@ -980,9 +980,9 @@ extern void		osAckRamromWrite(void);
 
 /* byte string operations */
 
-extern void     bcopy(const void *, void *, int);
-extern int      bcmp(const void *, const void *, int);
-extern void     bzero(void *, int);
+extern void     bcopy(const void *, void *, size_t);
+extern int      bcmp(const void *, const void *, size_t);
+extern void     bzero(void *, size_t);
 
 /* Miscellaneous operations */
